requirements.txt
transformers
torch
pandas
seaborn
numpy
tqdm
torchtext
torcheval
flask
line-bot-sdk
opencc
cnsenti